fix issues.

+++ app/core/src/main/java/com/thechance/core/data/usecase/market/DeleteMarketUseCase.kt
@@ -1,19 +1,25 @@
 package com.thechance.core.data.usecase.market
 
 import com.thechance.core.data.repository.HoneyMartRepository
+import com.thechance.core.data.utils.IdNotFoundException
 import com.thechance.core.data.utils.InvalidMarketIdException
 import com.thechance.core.data.utils.MarketDeletedException
-import com.thechance.core.data.utils.isValidId
+import com.thechance.core.data.utils.isInvalidId
 import org.koin.core.component.KoinComponent
 
 class DeleteMarketUseCase(private val repository: HoneyMartRepository) : KoinComponent {
     suspend operator fun invoke(marketId: Long?): Boolean {
-        return if (isValidId(marketId)) {
+        return if (isInvalidId(marketId)) {
             throw InvalidMarketIdException()
-        } else if (repository.isMarketDeleted(marketId!!)) {
-            throw MarketDeletedException()
         } else {
-            repository.deleteMarket(marketId)
+            val isDeleted = repository.isMarketDeleted(marketId!!)
+            if (isDeleted == null) {
+                throw IdNotFoundException()
+            } else if (isDeleted) {
+                throw MarketDeletedException()
+            } else {
+                repository.deleteMarket(marketId)
+            }
         }
     }
 }
